refactor(virtualmodel): split vmodel providers into openai and anthropic sub-packages
Previously a single Registry with a byProtocol map held both Anthropic and
OpenAI virtual models, and MockModel implemented both protocol sub-interfaces
on the same type. The result was that semantically OpenAI-only IDs (e.g.
virtual-gpt-4) could be invoked through /messages and vice versa, and the
"declares two protocols" lie at mock_model.go:361-370 was hard to police.

This change physically splits provider implementations into
internal/virtualmodel/anthropic and internal/virtualmodel/openai. Each
sub-package owns its own Registry, MockModel, scenario constructor, stream
events, and DefaultStream adapter; the root virtualmodel package retains only
the protocol-agnostic VirtualModel base interface and shared types.

Behavioral changes:
- A model registered for one protocol returns 404 (not 501) when called via
  the other protocol's endpoint — package membership is the protocol context.
- virtual-gpt-4 lives only in the OpenAI registry; virtual-claude-3 only in
  the Anthropic registry. echo-model and the tool-call mocks are registered
  in both as separate concrete instances.
- TransformModel and ClaudeCodeCompactTransform live in the anthropic
  sub-package and no longer carry a Protocols() method; the OpenAI registry
  cannot reference them.
- DelayProvider moves to the openai sub-package since it speaks OpenAI Chat
  wire format.

Updates: virtualserver.Service holds two registries and exposes
GetAnthropicRegistry / GetOpenAIRegistry. extension.RegisterVModelExtension
takes both registries and tags each item with a "provider" metadata key.
Tests are migrated into the sub-packages or updated to the new dual-registry
API; cross-protocol expectations now verify 404.
