+++ internal/extension/vmodel.go
@@ -4,11 +4,17 @@ import (
 	"fmt"
 
 	"github.com/tingly-dev/tingly-box/internal/virtualmodel"
+	anthropicvm "github.com/tingly-dev/tingly-box/internal/virtualmodel/anthropic"
+	openaivm "github.com/tingly-dev/tingly-box/internal/virtualmodel/openai"
 )
 
-// RegisterVModelExtension registers VModel as an extension in the extension registry
-func RegisterVModelExtension(extRegistry *ExtensionRegistry, vmRegistry *virtualmodel.Registry) error {
-	// Register VModel extension
+// RegisterVModelExtension registers VModel as an extension in the extension
+// registry, walking both per-provider virtual model registries. Each item is
+// tagged with a "provider" metadata key ("anthropic" or "openai") so the UI
+// can render the split. When the same model ID exists in both registries,
+// only the first one (anthropic) is registered as an item — the second is
+// skipped to avoid duplicate item IDs.
+func RegisterVModelExtension(extRegistry *ExtensionRegistry, anthropicReg *anthropicvm.Registry, openaiReg *openaivm.Registry) error {
 	vmodelExt := &Extension{
 		ID:          "vmodel",
 		Name:        "Virtual Models",
@@ -23,36 +29,41 @@ func RegisterVModelExtension(extRegistry *ExtensionRegistry, vmRegistry *virtual
 		return fmt.Errorf("failed to register vmodel extension: %w", err)
 	}
 
-	// Register all VirtualModels as ExtensionItems
-	virtualModels := vmRegistry.List()
-	for _, vm := range virtualModels {
-		item := &ExtensionItem{
-			ID:          vm.GetID(),
-			ExtensionID: "vmodel",
-			Name:        vm.GetName(),
-			Description: vm.GetDescription(),
-			Type:        string(vm.GetType()),
-			Metadata: map[string]interface{}{
-				"vmType": string(vm.GetType()),
-			},
-			Config: make(map[string]interface{}),
-		}
-
-		// Set icon based on type
-		switch vm.GetType() {
-		case virtualmodel.VirtualModelTypeStatic:
-			item.Icon = "message"
-		case virtualmodel.VirtualModelTypeProxy:
-			item.Icon = "compress"
-		case virtualmodel.VirtualModelTypeTool:
-			item.Icon = "build"
-		}
-
-		if err := extRegistry.RegisterItem(item); err != nil {
-			// Log but continue with other items
+	for _, vm := range anthropicReg.List() {
+		registerVModelItem(extRegistry, vm.GetID(), vm.GetName(), vm.GetDescription(), vm.GetType(), "anthropic")
+	}
+	for _, vm := range openaiReg.List() {
+		if extRegistry.GetItem("vmodel", vm.GetID()) != nil {
 			continue
 		}
+		registerVModelItem(extRegistry, vm.GetID(), vm.GetName(), vm.GetDescription(), vm.GetType(), "openai")
 	}
 
 	return nil
+}
+
+func registerVModelItem(extRegistry *ExtensionRegistry, id, name, description string, vmType virtualmodel.VirtualModelType, provider string) {
+	item := &ExtensionItem{
+		ID:          id,
+		ExtensionID: "vmodel",
+		Name:        name,
+		Description: description,
+		Type:        string(vmType),
+		Metadata: map[string]interface{}{
+			"vmType":   string(vmType),
+			"provider": provider,
+		},
+		Config: make(map[string]interface{}),
+	}
+
+	switch vmType {
+	case virtualmodel.VirtualModelTypeStatic:
+		item.Icon = "message"
+	case virtualmodel.VirtualModelTypeProxy:
+		item.Icon = "compress"
+	case virtualmodel.VirtualModelTypeTool:
+		item.Icon = "build"
+	}
+
+	_ = extRegistry.RegisterItem(item)
 }
