+++ internal/virtualmodel/anthropic/cc_compact.go
@@ -1,9 +1,9 @@
-package virtualmodel
+package anthropic
 
 import (
 	"strings"
 
-	"github.com/anthropics/anthropic-sdk-go"
+	sdk "github.com/anthropics/anthropic-sdk-go"
 	"github.com/sirupsen/logrus"
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
 	"github.com/tingly-dev/tingly-box/internal/smart_compact"
@@ -32,7 +32,7 @@ func (t *ClaudeCodeCompactTransform) Name() string {
 // Apply applies XML compression only if conditions are met.
 func (t *ClaudeCodeCompactTransform) Apply(ctx *transform.TransformContext) error {
 	switch req := ctx.Request.(type) {
-	case *anthropic.MessageNewParams:
+	case *sdk.MessageNewParams:
 		if !t.shouldApplyV1(req) {
 			logrus.Debugf("[claude_code_compact] v1: conditions not met, skipping")
 			return nil
@@ -40,7 +40,7 @@ func (t *ClaudeCodeCompactTransform) Apply(ctx *transform.TransformContext) erro
 		logrus.Debugf("[claude_code_compact] v1: conditions met, applying compression")
 		return t.inner.Apply(ctx)
 
-	case *anthropic.BetaMessageNewParams:
+	case *sdk.BetaMessageNewParams:
 		if !t.shouldApplyV1Beta(req) {
 			logrus.Debugf("[claude_code_compact] v1beta: conditions not met, skipping")
 			return nil
@@ -53,28 +53,21 @@ func (t *ClaudeCodeCompactTransform) Apply(ctx *transform.TransformContext) erro
 	}
 }
 
-// shouldApplyV1 checks v1 conditions for compression.
-func (t *ClaudeCodeCompactTransform) shouldApplyV1(req *anthropic.MessageNewParams) bool {
-	// Must have tools
-	if req.Tools == nil || len(req.Tools) == 0 {
+func (t *ClaudeCodeCompactTransform) shouldApplyV1(req *sdk.MessageNewParams) bool {
+	if len(req.Tools) == 0 {
 		return false
 	}
-	// Last user message must contain "compact"
 	return lastUserMessageContainsCompact(req.Messages)
 }
 
-// shouldApplyV1Beta checks v1beta conditions for compression.
-func (t *ClaudeCodeCompactTransform) shouldApplyV1Beta(req *anthropic.BetaMessageNewParams) bool {
-	// Must have tools
-	if req.Tools == nil || len(req.Tools) == 0 {
+func (t *ClaudeCodeCompactTransform) shouldApplyV1Beta(req *sdk.BetaMessageNewParams) bool {
+	if len(req.Tools) == 0 {
 		return false
 	}
-	// Last user message must contain "compact"
 	return lastUserMessageContainsCompactBeta(req.Messages)
 }
 
-// lastUserMessageContainsCompact checks if the last user message contains "compact".
-func lastUserMessageContainsCompact(messages []anthropic.MessageParam) bool {
+func lastUserMessageContainsCompact(messages []sdk.MessageParam) bool {
 	for i := len(messages) - 1; i >= 0; i-- {
 		if string(messages[i].Role) == "user" {
 			return extractAndCheckText(messages[i].Content)
@@ -83,8 +76,7 @@ func lastUserMessageContainsCompact(messages []anthropic.MessageParam) bool {
 	return false
 }
 
-// lastUserMessageContainsCompactBeta checks if the last user message contains "compact" for beta API.
-func lastUserMessageContainsCompactBeta(messages []anthropic.BetaMessageParam) bool {
+func lastUserMessageContainsCompactBeta(messages []sdk.BetaMessageParam) bool {
 	for i := len(messages) - 1; i >= 0; i-- {
 		if string(messages[i].Role) == "user" {
 			return extractAndCheckTextBeta(messages[i].Content)
@@ -93,8 +85,7 @@ func lastUserMessageContainsCompactBeta(messages []anthropic.BetaMessageParam) b
 	return false
 }
 
-// extractAndCheckText extracts text from v1 content blocks and checks for "compact".
-func extractAndCheckText(content []anthropic.ContentBlockParamUnion) bool {
+func extractAndCheckText(content []sdk.ContentBlockParamUnion) bool {
 	for _, block := range content {
 		if block.OfText != nil {
 			if strings.Contains(strings.ToLower(block.OfText.Text), "compact") {
@@ -105,8 +96,7 @@ func extractAndCheckText(content []anthropic.ContentBlockParamUnion) bool {
 	return false
 }
 
-// extractAndCheckTextBeta extracts text from beta content blocks and checks for "compact".
-func extractAndCheckTextBeta(content []anthropic.BetaContentBlockParamUnion) bool {
+func extractAndCheckTextBeta(content []sdk.BetaContentBlockParamUnion) bool {
 	for _, block := range content {
 		if block.OfText != nil {
 			if strings.Contains(strings.ToLower(block.OfText.Text), "compact") {
