+++ internal/virtualmodel/anthropic/cc_compact_test.go
@@ -1,67 +1,36 @@
-package virtualmodel
+package anthropic
 
 import (
 	"strings"
 	"testing"
 
-	"github.com/anthropics/anthropic-sdk-go"
+	sdk "github.com/anthropics/anthropic-sdk-go"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
 	"github.com/tingly-dev/tingly-box/internal/smart_compact"
 )
 
-// TestClaudeCodeCompact_Compression tests that claude-code-compact actually compresses messages
+// TestClaudeCodeCompact_Compression tests that claude-code-compact actually compresses messages.
 func TestClaudeCodeCompact_Compression(t *testing.T) {
 	vm := newCompactVM()
 
-	// Create a request with multiple rounds and tools + compact command
-	originalMessages := []anthropic.BetaMessageParam{
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("First user message"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleAssistant,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("First assistant response"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Second user message"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleAssistant,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Second assistant response"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("<command>compact</command>"),
-			},
-		},
+	originalMessages := []sdk.BetaMessageParam{
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("First user message")}},
+		{Role: sdk.BetaMessageParamRoleAssistant, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("First assistant response")}},
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Second user message")}},
+		{Role: sdk.BetaMessageParamRoleAssistant, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Second assistant response")}},
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("<command>compact</command>")}},
 	}
 
 	req := &protocol.AnthropicBetaMessagesRequest{
-		BetaMessageNewParams: anthropic.BetaMessageNewParams{
+		BetaMessageNewParams: sdk.BetaMessageNewParams{
 			Messages: originalMessages,
-			Tools: []anthropic.BetaToolUnionParam{
-				{
-					OfTool: &anthropic.BetaToolParam{
-						Name: "read_file",
-					},
-				},
+			Tools: []sdk.BetaToolUnionParam{
+				{OfTool: &sdk.BetaToolParam{Name: "read_file"}},
 			},
 		},
 	}
 
-	// Process the request
 	result, err := vm.HandleAnthropic(req)
 	if err != nil {
 		t.Fatalf("HandleAnthropic failed: %v", err)
@@ -71,19 +40,16 @@ func TestClaudeCodeCompact_Compression(t *testing.T) {
 	t.Logf("Result message count: %d", len(req.Messages))
 	t.Logf("Result content: %s", extractTextFromVModelResponse(result))
 
-	// Verify messages were compressed (should be fewer than original)
 	if len(req.Messages) >= len(originalMessages) {
 		t.Errorf("Expected compressed messages (%d < %d), got %d",
 			len(req.Messages), len(originalMessages), len(req.Messages))
 	}
 
-	// Verify the result contains a compressed message
 	hasCompressedContent := false
 	for _, msg := range req.Messages {
 		for _, block := range msg.Content {
 			if block.OfText != nil && len(block.OfText.Text) > 0 {
 				content := block.OfText.Text
-				// Compressed content should be XML format
 				if len(content) > 100 && (strings.Contains(content, "<conversation>") || strings.Contains(content, "<user>")) {
 					hasCompressedContent = true
 					t.Logf("Found compressed content (length: %d): %s", len(content), truncate(content, 200))
@@ -96,54 +62,34 @@ func TestClaudeCodeCompact_Compression(t *testing.T) {
 		t.Error("Expected to find compressed content in result messages, but none found")
 	}
 
-	// Verify response content - it should be the compressed summary
 	contentText := extractTextFromVModelResponse(result)
 	if contentText == "" {
 		t.Error("Expected non-empty content (compressed summary)")
 	}
 
-	// Content should contain analysis or summary markers
 	if !strings.Contains(contentText, "<analysis>") && !strings.Contains(contentText, "<summary>") {
 		t.Errorf("Expected content to contain compressed summary markers, got: %s", truncate(contentText, 200))
 	}
 
-	// Verify stop reason
 	if result.StopReason != "end_turn" {
 		t.Errorf("Expected stop_reason 'end_turn', got %q", result.StopReason)
 	}
-
-	t.Log("✓ Claude Code Compact compression test passed")
 }
 
-// TestClaudeCodeCompact_NoCompressionWithoutCommand tests that compression doesn't happen without <command>compact</command>
+// TestClaudeCodeCompact_NoCompressionWithoutCommand tests that compression doesn't happen without <command>compact</command>.
 func TestClaudeCodeCompact_NoCompressionWithoutCommand(t *testing.T) {
 	vm := newCompactVM()
 
-	// Request WITHOUT compact command
-	originalMessages := []anthropic.BetaMessageParam{
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Regular user message"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleAssistant,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Regular assistant response"),
-			},
-		},
+	originalMessages := []sdk.BetaMessageParam{
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Regular user message")}},
+		{Role: sdk.BetaMessageParamRoleAssistant, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Regular assistant response")}},
 	}
 
 	req := &protocol.AnthropicBetaMessagesRequest{
-		BetaMessageNewParams: anthropic.BetaMessageNewParams{
+		BetaMessageNewParams: sdk.BetaMessageNewParams{
 			Messages: originalMessages,
-			Tools: []anthropic.BetaToolUnionParam{
-				{
-					OfTool: &anthropic.BetaToolParam{
-						Name: "read_file",
-					},
-				},
+			Tools: []sdk.BetaToolUnionParam{
+				{OfTool: &sdk.BetaToolParam{Name: "read_file"}},
 			},
 		},
 	}
@@ -153,16 +99,11 @@ func TestClaudeCodeCompact_NoCompressionWithoutCommand(t *testing.T) {
 		t.Fatalf("HandleAnthropic failed: %v", err)
 	}
 
-	t.Logf("Original message count: %d", len(originalMessages))
-	t.Logf("Result message count: %d", len(req.Messages))
-
-	// Without compact command, messages should NOT be compressed
 	if len(req.Messages) != len(originalMessages) {
 		t.Errorf("Without compact command, expected message count to remain %d, got %d",
 			len(originalMessages), len(req.Messages))
 	}
 
-	// Verify no compressed content
 	for _, msg := range req.Messages {
 		for _, block := range msg.Content {
 			if block.OfText != nil {
@@ -173,43 +114,24 @@ func TestClaudeCodeCompact_NoCompressionWithoutCommand(t *testing.T) {
 			}
 		}
 	}
-
-	t.Log("✓ No compression without command test passed")
 }
 
-// TestClaudeCodeCompact_NoCompressionWithoutTools tests that compression doesn't happen without tools
+// TestClaudeCodeCompact_NoCompressionWithoutTools tests that compression doesn't happen without tools.
 func TestClaudeCodeCompact_NoCompressionWithoutTools(t *testing.T) {
 	vm := NewTransformModel(&TransformModelConfig{
 		ID:    "claude-code-compact",
 		Chain: transform.NewTransformChain([]transform.Transform{smart_compact.NewXMLCompactTransform()}),
 	})
 
-	// Request with compact command but NO tools
-	originalMessages := []anthropic.BetaMessageParam{
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Message 1"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleAssistant,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("Response 1"),
-			},
-		},
-		{
-			Role: anthropic.BetaMessageParamRoleUser,
-			Content: []anthropic.BetaContentBlockParamUnion{
-				anthropic.NewBetaTextBlock("<command>compact</command>"),
-			},
-		},
+	originalMessages := []sdk.BetaMessageParam{
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Message 1")}},
+		{Role: sdk.BetaMessageParamRoleAssistant, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("Response 1")}},
+		{Role: sdk.BetaMessageParamRoleUser, Content: []sdk.BetaContentBlockParamUnion{sdk.NewBetaTextBlock("<command>compact</command>")}},
 	}
 
 	req := &protocol.AnthropicBetaMessagesRequest{
-		BetaMessageNewParams: anthropic.BetaMessageNewParams{
+		BetaMessageNewParams: sdk.BetaMessageNewParams{
 			Messages: originalMessages,
-			// No tools!
 		},
 	}
 
@@ -218,16 +140,12 @@ func TestClaudeCodeCompact_NoCompressionWithoutTools(t *testing.T) {
 		t.Fatalf("HandleAnthropic failed: %v", err)
 	}
 
-	// Without tools, compression should NOT happen (per claude-code-compact logic)
 	if len(req.Messages) != len(originalMessages) {
 		t.Logf("Without tools, message count changed from %d to %d (compression may still occur depending on implementation)",
 			len(originalMessages), len(req.Messages))
 	}
-
-	t.Log("✓ No compression without tools test passed")
 }
 
-// Helper functions
 func newCompactVM() *TransformModel {
 	return NewTransformModel(&TransformModelConfig{
 		ID:    "claude-code-compact",
@@ -242,7 +160,6 @@ func truncate(s string, maxLen int) string {
 	return s[:maxLen] + "..."
 }
 
-// extractTextFromVModelResponse extracts text content from VModelResponse
 func extractTextFromVModelResponse(resp VModelResponse) string {
 	var text string
 	for _, block := range resp.Content {
