+++ internal/virtualmodel/anthropic/mock_model.go
@@ -1,0 +1,153 @@
+package anthropic
+
+import (
+	"encoding/json"
+	"time"
+
+	sdk "github.com/anthropics/anthropic-sdk-go"
+	"github.com/tingly-dev/tingly-box/internal/protocol"
+	"github.com/tingly-dev/tingly-box/internal/protocol/token"
+	"github.com/tingly-dev/tingly-box/internal/virtualmodel"
+)
+
+// MockModelConfig holds configuration for an Anthropic mock virtual model
+// (static or tool).
+type MockModelConfig struct {
+	ID           string
+	Name         string
+	Description  string
+	Content      string        // static: fixed response text
+	StopReason   string        // default: "stop" (or "tool_use" if ToolCall is set)
+	Delay        time.Duration // simulated latency
+	StreamChunks []string      // optional custom stream chunks
+
+	// tool-type: if set, the response includes a tool_use block.
+	ToolCall *virtualmodel.ToolCallConfig
+}
+
+// MockModel is an Anthropic-only mock virtual model. It returns a fixed
+// content block (or tool_use block) and supports streaming via per-chunk
+// token splitting.
+type MockModel struct {
+	cfg *MockModelConfig
+}
+
+// Compile-time interface check.
+var _ VirtualModel = (*MockModel)(nil)
+
+// NewMockModel creates a MockModel. StopReason defaults to "stop"
+// (or "tool_use" if ToolCall is set).
+func NewMockModel(cfg *MockModelConfig) *MockModel {
+	if cfg.StopReason == "" {
+		if cfg.ToolCall != nil {
+			cfg.StopReason = "tool_use"
+		} else {
+			cfg.StopReason = "stop"
+		}
+	}
+	return &MockModel{cfg: cfg}
+}
+
+func (m *MockModel) GetID() string { return m.cfg.ID }
+
+func (m *MockModel) GetName() string { return m.cfg.Name }
+
+func (m *MockModel) GetDescription() string {
+	if m.cfg.Description != "" {
+		return m.cfg.Description
+	}
+	return "A virtual model that returns fixed responses for testing"
+}
+
+func (m *MockModel) GetType() virtualmodel.VirtualModelType {
+	if m.cfg.ToolCall != nil {
+		return virtualmodel.VirtualModelTypeTool
+	}
+	return virtualmodel.VirtualModelTypeStatic
+}
+
+func (m *MockModel) SimulatedDelay() time.Duration { return m.cfg.Delay }
+
+func (m *MockModel) ToModel() virtualmodel.Model {
+	return virtualmodel.Model{
+		ID:      m.cfg.ID,
+		Object:  "model",
+		Created: 0,
+		OwnedBy: "tingly-box-virtual",
+	}
+}
+
+func (m *MockModel) streamChunks() []string {
+	if len(m.cfg.StreamChunks) > 0 {
+		return m.cfg.StreamChunks
+	}
+	return token.SplitIntoChunks(m.cfg.Content)
+}
+
+// HandleAnthropic returns fixed content from config in Anthropic format.
+func (m *MockModel) HandleAnthropic(_ *protocol.AnthropicBetaMessagesRequest) (VModelResponse, error) {
+	if m.cfg.ToolCall != nil {
+		return m.toolResponse(), nil
+	}
+	return m.staticResponse(), nil
+}
+
+func (m *MockModel) staticResponse() VModelResponse {
+	return VModelResponse{
+		Content: []sdk.BetaContentBlockParamUnion{
+			{OfText: &sdk.BetaTextBlockParam{Text: m.cfg.Content}},
+		},
+		StopReason: sdk.BetaStopReason(m.cfg.StopReason),
+	}
+}
+
+func (m *MockModel) toolResponse() VModelResponse {
+	tc := m.cfg.ToolCall
+	displayText := virtualmodel.ToolCallDisplayContent(tc.Arguments)
+	inputJSON, _ := json.Marshal(tc.Arguments)
+
+	return VModelResponse{
+		Content: []sdk.BetaContentBlockParamUnion{
+			{OfText: &sdk.BetaTextBlockParam{Text: displayText}},
+			{OfToolUse: &sdk.BetaToolUseBlockParam{
+				ID:    "toolu_virtual",
+				Name:  tc.Name,
+				Input: json.RawMessage(inputJSON),
+			}},
+		},
+		StopReason: "tool_use",
+	}
+}
+
+// HandleAnthropicStream streams fixed content using configured chunks with simulated delay.
+func (m *MockModel) HandleAnthropicStream(req *protocol.AnthropicBetaMessagesRequest, emit func(any)) error {
+	resp, err := m.HandleAnthropic(req)
+	if err != nil {
+		return err
+	}
+	emit(StreamStartEvent{MsgID: "msg_virtual", Model: m.cfg.ID})
+	delay := m.cfg.Delay
+	chunks := m.streamChunks()
+	for i, blk := range resp.Content {
+		if blk.OfText != nil {
+			for _, chunk := range chunks {
+				if delay > 0 {
+					time.Sleep(delay / time.Duration(len(chunks)))
+				} else {
+					time.Sleep(50 * time.Millisecond)
+				}
+				emit(TextDeltaEvent{Index: i, Text: chunk})
+			}
+		} else if blk.OfToolUse != nil {
+			inputJSON, _ := json.Marshal(blk.OfToolUse.Input)
+			emit(ToolUseEvent{
+				Index: i,
+				ID:    blk.OfToolUse.ID,
+				Name:  blk.OfToolUse.Name,
+				Input: inputJSON,
+			})
+		}
+	}
+	emit(DoneEvent{StopReason: string(resp.StopReason)})
+	return nil
+}
