+++ internal/virtualmodel/anthropic/transform_model.go
@@ -1,12 +1,13 @@
-package virtualmodel
+package anthropic
 
 import (
 	"fmt"
 	"time"
 
-	anthropic "github.com/anthropics/anthropic-sdk-go"
+	sdk "github.com/anthropics/anthropic-sdk-go"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
+	"github.com/tingly-dev/tingly-box/internal/virtualmodel"
 )
 
 // TransformModelConfig holds configuration for a transform virtual model (proxy).
@@ -18,15 +19,15 @@ type TransformModelConfig struct {
 	Chain       *transform.TransformChain // applied first
 }
 
-// TransformModel implements AnthropicVirtualModel for proxy/transform virtual models.
-// HandleAnthropic applies Chain and Transformer in-place on req; ResponseAnthropic reads the result.
-// It does NOT implement OpenAIChatVirtualModel — transform models are Anthropic-only.
+// TransformModel is an Anthropic-protocol proxy model. HandleAnthropic
+// applies Chain and Transformer in-place on req and returns the resulting
+// last-message text.
 type TransformModel struct {
 	cfg *TransformModelConfig
 }
 
 // Compile-time interface check.
-var _ AnthropicVirtualModel = (*TransformModel)(nil)
+var _ VirtualModel = (*TransformModel)(nil)
 
 // NewTransformModel creates a TransformModel.
 func NewTransformModel(cfg *TransformModelConfig) *TransformModel {
@@ -39,18 +40,15 @@ func (m *TransformModel) GetName() string { return m.cfg.Name }
 
 func (m *TransformModel) GetDescription() string { return m.cfg.Description }
 
-func (m *TransformModel) GetType() VirtualModelType { return VirtualModelTypeProxy }
+func (m *TransformModel) GetType() virtualmodel.VirtualModelType {
+	return virtualmodel.VirtualModelTypeProxy
+}
 
 // SimulatedDelay is always 0 — transform models don't simulate latency.
 func (m *TransformModel) SimulatedDelay() time.Duration { return 0 }
 
-// Protocols declares that TransformModel is Anthropic-only.
-func (m *TransformModel) Protocols() []protocol.APIType {
-	return []protocol.APIType{protocol.TypeAnthropicBeta}
-}
-
-func (m *TransformModel) ToModel() Model {
-	return Model{
+func (m *TransformModel) ToModel() virtualmodel.Model {
+	return virtualmodel.Model{
 		ID:      m.cfg.ID,
 		Object:  "model",
 		Created: 0,
@@ -58,23 +56,21 @@ func (m *TransformModel) ToModel() Model {
 	}
 }
 
-// HandleAnthropicStream delegates to DefaultAnthropicStream since TransformModel is batch-only.
+// HandleAnthropicStream delegates to DefaultStream since TransformModel is batch-only.
 func (m *TransformModel) HandleAnthropicStream(req *protocol.AnthropicBetaMessagesRequest, emit func(any)) error {
-	return DefaultAnthropicStream(m, req, emit)
+	return DefaultStream(m, req, emit)
 }
 
 // HandleAnthropic applies Chain then Transformer to req in-place and returns the response.
 func (m *TransformModel) HandleAnthropic(req *protocol.AnthropicBetaMessagesRequest) (VModelResponse, error) {
 	ctx := transform.NewTransformContext(&req.BetaMessageNewParams)
 
-	// Apply Chain first
 	if m.cfg.Chain != nil {
 		if _, err := m.cfg.Chain.Execute(ctx); err != nil {
 			return VModelResponse{}, fmt.Errorf("transform chain failed: %w", err)
 		}
 	}
 
-	// Apply Transformer
 	if m.cfg.Transformer != nil {
 		if err := m.cfg.Transformer.Apply(ctx); err != nil {
 			return VModelResponse{}, fmt.Errorf("transformer failed: %w", err)
@@ -91,7 +87,7 @@ func (m *TransformModel) HandleAnthropic(req *protocol.AnthropicBetaMessagesRequ
 		}
 	}
 	return VModelResponse{
-		Content:    []anthropic.BetaContentBlockParamUnion{{OfText: &anthropic.BetaTextBlockParam{Text: text}}},
+		Content:    []sdk.BetaContentBlockParamUnion{{OfText: &sdk.BetaTextBlockParam{Text: text}}},
 		StopReason: "end_turn",
 	}, nil
 }
